Remove batching and yielded laziness
The structure that was put in place for laziness and batching support
made it impossible to extract cache_key's out of the yielded object.
Typically the yielded object would be a serializer or decorator which
may have its own cache key. The previous strategy would never use that
cache key, making it impractical to use a custom cache key.

As of this commit batching functionality has also been removed.